+++ web/src/components/loading/Loading.tsx
@@ -3,12 +3,12 @@ import { LoadingOutlined } from '@ant-design/icons';
 import { Col, Row, Space, Spin } from 'antd';
 
 interface LoadingProps {
-    message: string    
+    message: string
 }
 
 export const Loading = (props: LoadingProps) => {
     const antIcon = <LoadingOutlined style={{ fontSize: 60 }} spin />;
-    return <div>
+    return <div style={{ background: 'white' }}>
         <Col span={8} offset={8} >
             <div className="loading_content">
                 <Spin size="large" indicator={antIcon} />
